feat: container, 스타일 파일 분리, nav 컴포넌트 분리
